app/app/requirements.txt
flask==3.0.0
Flask-Session==0.6.0
ldap3==2.9.1
openai==1.3.5
pandas==2.1.1
numpy==1.26.1
plotly==5.18.0
msal==1.26.0
gunicorn==21.2.0
snowflake-connector-python==3.7.0
